feat: Move decomposition default groups to new file
