+++ py/torch_tensorrt/dynamo/_defaults.py
@@ -10,6 +10,5 @@ VERSION_COMPATIBLE = False
 OPTIMIZATION_LEVEL = None
 TRUNCATE_LONG_AND_DOUBLE = False
 USE_PYTHON_RUNTIME = False
-
 USE_FAST_PARTITIONER = True
 ENABLE_EXPERIMENTAL_DECOMPOSITIONS = False
